fix(debug): 移除 debug:impl 多余的 Kuikly/Compose 插件和 KSP 依赖
debug:impl 是纯业务逻辑模块，不需要 Kuikly/Compose 插件。
之前 17695b79 误添加了这些插件，现在移除。


fix(ohos): 修复 KNOI 处理器类型不支持和编译错误

- 添加 kotlinx-coroutines-core 依赖（KBA 版本）
- 修复 KNOI 不支持 Any/Float 类型的问题（requestPermissions 移除 context 参数，getScreenDensity 返回 Double）
- ArrayBuffer 改用 Base64 字符串传输二进制数据
- 修复 NtpClock 非局部 return 语法错误
- 补全 capturePhoto/recordVideo/pickMedia 接口声明

+++ common/base/src/ohosArm64Main/kotlin/com/kuikly/init/common/base/platform/screen/ScreenInfo.kt
@@ -14,8 +14,8 @@ actual fun provideScreenInfo(): ScreenInfo {
             widthPx = service?.getScreenWidth() ?: 0,
             heightPx = service?.getScreenHeight() ?: 0,
             densityDpi = service?.getScreenDensityDpi() ?: 160,
-            density = service?.getScreenDensity() ?: 1.0f,
-            scaledDensity = service?.getScreenDensity() ?: 1.0f,
+            density = (service?.getScreenDensity() ?: 1.0).toFloat(),
+            scaledDensity = (service?.getScreenDensity() ?: 1.0).toFloat(),
             rotation = service?.getScreenRotation() ?: 0
         )
     } catch (e: Exception) {
